requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
DateTime==5.1
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
openai==0.27.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
zope.interface==6.0
